Normal Skeletons no longer spawn in nether, some wither skeles have bows
